style(open-api-page): continue cleanup, switch to flight icons
