+++ packages/open-api-page/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@apidevtools/json-schema-ref-parser": "^9.0.9",
+    "@hashicorp/flight-icons": "^1.1.0",
     "@hashicorp/platform-markdown-utils": "^0.1.3",
     "@hashicorp/react-content": "^8.0.2",
     "change-case": "^4.1.2",
