+++ packages/open-api-page/partials/operation-object/style.module.css
@@ -1,20 +1,33 @@
 .root {
-  border-bottom: 1px solid var(--gray-5);
+  --border-bottom-color: var(--gray-5);
+  --header-title-color: var(--gray-2);
+  --header-text-color: var(--gray-3);
+
+  border-bottom: 1px solid var(--border-bottom-color);
   margin: 0;
-  margin-top: -1px;
+
+  /* margin-top: -1px; */
   position: relative;
   transition: border-color 0.2s;
 
-  &[data-is-hovered='true'] {
-    border-bottom-color: var(--brand);
-    z-index: 1;
+  &.isHeaderHovered {
+    --border-bottom-color: var(--brand);
+    --header-title-color: var(--brand-link);
+    --header-text-color: var(--brand-link);
   }
 }
 
 .header {
+  background: none;
+  border: none;
   display: flex;
-  position: relative;
+  font-family: inherit;
+  font-size: inherit;
+  margin: 0;
   padding: 16px 0;
+  position: relative;
+  text-align: left;
+  width: 100%;
 
   &:hover {
     cursor: pointer;
@@ -22,6 +35,7 @@
 }
 
 .meta {
+  display: block;
   width: 100%;
 }
 
@@ -29,38 +43,17 @@
   position: relative;
   display: inline-block;
   composes: g-type-body-x-strong from global;
-  color: var(--gray-2);
+  color: var(--header-title-color);
   margin: 0 0 4px 0;
   padding: 0;
-
-  &[data-is-highlighted='true'] {
-    &::before {
-      content: '\200B';
-      position: absolute;
-      top: -4px;
-      left: -8px;
-      right: -8px;
-      bottom: -4px;
-      background: var(--warning-l1);
-      z-index: -1;
-      border-radius: 8px;
-    }
-  }
-
-  &[data-is-hovered='true'] {
-    color: var(--brand-link);
-  }
 }
 
 .endpoint {
   composes: g-type-code from global;
+  display: block;
 
   & .method {
-    color: var(--gray-3);
-
-    &[data-is-hovered='true'] {
-      color: var(--brand-link);
-    }
+    color: var(--header-text-color);
   }
 
   & .path {
@@ -69,13 +62,12 @@
   }
 }
 
-.toggleButton {
-  border: none;
+.toggleText {
   color: var(--brand-link);
   background: none;
   display: flex;
+  align-self: stretch;
   align-items: center;
-  font-size: inherit;
   margin: 0;
   padding: 0 0 0 24px;
   transition: color 0.2s;
@@ -92,22 +84,9 @@
     align-items: center;
     justify-content: center;
 
-    & svg {
-      & [fill] {
-        fill: var(--brand-link);
-      }
-      & [stroke] {
-        stroke: var(--brand-link);
-      }
-    }
-
     &.isCollapsed {
       transform: rotate(0deg);
     }
-  }
-
-  &[data-is-hovered='true'] {
-    color: var(--brand);
   }
 }
 
@@ -144,18 +123,4 @@
   }
 
   margin: 8px 0 24px 0;
-}
-
-.twoColumnLayout {
-  display: flex;
-
-  & > div:nth-child(1),
-  & > div:nth-child(3) {
-    position: relative;
-    width: calc(50% - 16px);
-  }
-
-  & > div:nth-child(2) {
-    width: 32px;
-  }
 }
