+++ packages/open-api-page/partials/property-object/style.module.css
@@ -35,15 +35,6 @@
       align-items: center;
       justify-content: center;
 
-      & svg {
-        & [fill] {
-          fill: var(--brand-link);
-        }
-        & [stroke] {
-          stroke: var(--brand-link);
-        }
-      }
-
       &.isCollapsed {
         transform: rotate(0deg);
       }
