+++ packages/open-api-page/partials/two-column-layout/index.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import s from './style.module.css'
+
+function TwoColumnLayout({ columnOne, columnTwo }) {
+  return (
+    <div className={s.twoColumnLayout}>
+      <div>{columnOne}</div>
+      <div></div>
+      <div>{columnTwo}</div>
+    </div>
+  )
+}
+
+export default TwoColumnLayout
